+++ android/app/build.gradle
@@ -45,8 +45,8 @@ android {
         targetSdkVersion 30
 //        versionCode flutterVersionCode.toInteger()
 //        versionName flutterVersionName
-        versionCode 4
-        versionName "2.4.0"
+        versionCode 5
+        versionName "2.4.1"
     }
 
     signingConfigs {
